requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
python-multipart==0.0.12
python-dotenv==1.0.1
requests==2.32.3
Pillow==11.0.0
jinja2==3.1.4